+++ docs/doxygen/html/search/all_e.js
@@ -7,7 +7,7 @@ var searchData=
   ['r_5fq_5fp01_5fboundaries_4',['R_Q_P01_boundaries',['../qnorm___r_8h.html#a40b69d5b59cec96b4d7d55e130c15ee1',1,'qnorm_R.h']]],
   ['rcppexports_2ecpp_5',['RcppExports.cpp',['../_rcpp_exports_8cpp.html',1,'']]],
   ['rdf_6',['rdf',['../classfastglmm_lib_1_1_model_fit.html#a2c3704281b2ab5fcd244a4cc4da67386',1,'fastglmmLib::ModelFit']]],
-  ['residuals_7',['residuals',['../structfastglmm_lib_1_1_l_m_work.html#a43f0f83be99b847dd068501a08e0603d',1,'fastglmmLib::LMWork::residuals'],['../classfastglmm_lib_1_1_model_fit.html#a2d6e59e0f7fab81d35e413ff3f7c863d',1,'fastglmmLib::ModelFit::residuals'],['../classfastglmm_lib_1_1fastglmm.html#ae312a6a0452c8097e263387210275451',1,'fastglmmLib::fastglmm::residuals()'],['../classfastglmm_lib_1_1fastlmm.html#a4c23fd36dd495d027db4a9752c4a2adb',1,'fastglmmLib::fastlmm::residuals()']]],
+  ['residuals_7',['residuals',['../structfastglmm_lib_1_1_l_m_work.html#a43f0f83be99b847dd068501a08e0603d',1,'fastglmmLib::LMWork::residuals'],['../classfastglmm_lib_1_1_model_fit.html#a2d6e59e0f7fab81d35e413ff3f7c863d',1,'fastglmmLib::ModelFit::residuals'],['../classfastglmm_lib_1_1fastglmm.html#ae312a6a0452c8097e263387210275451',1,'fastglmmLib::fastglmm::residuals()'],['../classfastglmm_lib_1_1fastlmm.html#a8e7e12e71a603b8595318675cef875b3',1,'fastglmmLib::fastlmm::residuals()']]],
   ['residuals_5fpearson_8',['residuals_pearson',['../classfastglmm_lib_1_1fastglmm.html#a0d841e00a3764bb9f67dd228ce388cf6',1,'fastglmmLib::fastglmm']]],
   ['reweight_9',['reweight',['../classfastglmm_lib_1_1spectral_decomp.html#a147f7b2391dfc9a0c3677d39c07b0aa6',1,'fastglmmLib::spectralDecomp']]],
   ['ru_10',['ru',['../classfastglmm_lib_1_1_model_fit_l_m_m.html#a26571524a40d01861291d366798c6a29',1,'fastglmmLib::ModelFitLMM']]]